SETI-376: Perform unit tests when building RPM package
There is a time period between running check pipeline and building
the package. If we run unit tests only in check pipeline, a
dependency could unexpectedly update and we would package zuul
with different and untested code.

Let's run unit tests for zuul immediately after building it and
with the same versions of dependencies.