TRIVIAL: Assert how many API calls are done during tests
The API calls made to GitHub were counted in python session with github3
for all calls like this:
  > import github3
  > gh = github3.login(token='mytoken')
  > rl_start = gh.ratelimit_remaining
  > gh.issue('hrubi', 'ci-infra', 134).create_comment('another comment')
  > rl_end = gh.ratelimit_remaining
  > rl_start - rl_end
  2

I believe all the steps which do two API calls can be made to do only a
single one when using the github API directly instead of github3.py.

Also getting the filenames from the GithubSource is very uneconomic,
because every configured pipeline will get the change again and issue
this API call.

Note that the API call counting is dumb for getUser, as in real it
creates a lazy object, which hits github only when needed.