SETI-1323: Reduce API calls in getChange
During the lifetime of a github event, getChange is called several
times, multiplied by the number of configured pipelines. It makes sense
to make the API calls to GitHub upfront on the event retrieval, instead
on each getChange call. That way all the information is going to be
available in the event, which is passed to getChange.

Alternative (similar as in GerritConnection) would be to build a cache of
the PRs on the GithubConnection.