SETI-377: Provide an option for reporting on build abort
Sometimes we need to report that a build in some pipeline was
aborted. Create a new type of reporter called 'abort' which is
triggered when a build is canceled.

Example configuration:
A developer force-pushes after gate pipeline is triggered and
corresponding build is aborted. In this case, we want to remove
'gate' label and set gate pipeline status to 'failure'.