SETI-1767: github reporter: report status and labels before merge
As part of SETI-377, order of reporting items to github PR was changed
to report failure more easily in case of merge failure.
Particularly, reported was supposed to merge the PR and then change
the pipeline status. This does not work on protected branches:
required status needs to be reported first and then merging is allowed
by Github.

Let's change the order of reporting, so that reporter reports status and
labels first and then merge the PR. Comment reporting happens after
merging PR to prevent reporting 2 comments (first "Merge successful",
second "Merge failed") which would be confusing.