SETI-1829: test zuul-cloner corner case
If zuul-cloner is aborted in the middle, it can leave pointing the
repository's HEAD pointing to refs/remotes/origin/master (or whatever
the remote head is). If subsequent invocation of zuul-cloner wants to
work with the `git branch` command - eg. delete stale remote tracking
branches, git will fail with

    git.exc.GitCommandError: Cmd('git') failed due to: exit code(128)
      cmdline: git branch -d -r origin/extra_branch
      stderr: 'fatal: HEAD not found below refs/heads!'

Note that zuul-cloner calling the zuul.merger.merger.Repo.reset() is
IMO rather accidental. The reset to remote HEAD is perhaps useful for
the zuul-merger, but not for the cloner.