SETI-1829: do not create symbolic refs in cloner/merger
`git branch` command freaks out when HEAD points to something out of
refs/heads. This was the case when calling the zuul.merger.merger.Repo.prune() method.

Fix this by not creating the HEAD a symbolic ref to remote ref, but
rather make it an ordinary ref to a commit to which the remote ref
points to.