Revert "Fix git prune order in zuul-cloner"
To fully fix the test introduced for simulating zuul-cloner being
aborted in the middle, the repo.reset() has to run before repo.prune()
to get the repository out of the state when HEAD points to a remote ref.

Note that in fact the situation provoked in the test should not happen
anymore, because the HEAD is no longer a symbolic ref. However, this
commit can get some repo which is already "broken" to a working state.

We also get rid of an unnecessary comment which is no longer true.

This reverts commit f5e16815546fa6d9e7edbc8b41a4e5ae68d12378.